--- src/App.css
@@ -1,32 +1,0 @@
-.authentication {
-  padding: 15px;
-  position: relative;
-  font-weight: normal;
-  font-family: "Open Sans",
-    "Helvetica Neue",
-    Helvetica,
-    Arial,
-    sans-serif;
-  font-size: 32px;
-  text-align: center;
-  color: #333;
-}
-
-.authentication img {
-  position: relative;
-  border-radius: 50%;
-  top: 4px;
-  padding-right: 15px;
-  width: 32px;
-  height: 32px;
-}
-
-.authentication button {
-  position: relative;
-  top: -5px;
-
-}
-
-span.welcome {
-  padding-right: 15px;
-}
