+++ src/Header.js
@@ -8,9 +8,6 @@ const Header = () => {
       <img src="//cdn.jsdelivr.net/emojione/assets/png/1f638.png" width="32" height="32" alt="" />
       <Link to="/">Emoji Search</Link>
       <img src="//cdn.jsdelivr.net/emojione/assets/png/1f63a.png" width="32" height="32" alt="" />
-      <div className="Menu">
-        <Link to="/favorites">Favorites</Link>
-      </div>
     </header>
   );
 };
